requirements.txt
torch
torchvision
fastapi
uvicorn
Pillow
requests
python-multipart